test(#146) : 분리
